Fix the Wi32 compiler warnings.

+++ src/scripting/BasicScriptObject.cpp
@@ -1285,7 +1285,7 @@ bool BasicScriptObject::ScintillaCommandsDispId(wchar_t * name, DISPID& id)
 
 HRESULT BasicScriptObject::ScintillaCommandInvoke(DISPID id, WORD flags, DISPPARAMS* args, VARIANT* ret)
 {
-    int vectIndex = id - 1000;
+    size_t vectIndex = id - 1000;
     if ((vectIndex > 0) && (vectIndex < m_ScintillaCmds.size()))
     {
         _variant_t p1;
@@ -1383,4 +1383,3 @@ HRESULT BasicScriptObject::ScintillaCommandInvoke(DISPID id, WORD flags, DISPPAR
     }
     return DISP_E_MEMBERNOTFOUND;
 }
-
